requirements.txt
streamlit>=1.30.0
pandas>=2.0.3
numpy>=1.25.0
matplotlib>=3.7.1
plotly>=5.18.0
openai>=1.32.0
python-dotenv>=1.0.1
